before i X2

+++ draw_bounding_box/video/video.py
@@ -8,7 +8,7 @@ import cv2
 from PyQt5.QtCore import QTimer, QRect
 from PyQt5.QtCore import Qt
 
-
+import string
 
 global img
 global point1, point2
@@ -127,7 +127,7 @@ class Video(QWidget):
                                )
         self.label_num.setFixedSize(900, 50)  # width height
         self.label_num.move(50, 0)
-        self.label_num.setStyleSheet("QLabel{background:aqua;}")
+        self.label_num.setStyleSheet("QLabel{background:lightsteelblue;}")
 
         #播放器滑动条
         # 设置最大最小值
@@ -203,6 +203,17 @@ class Video(QWidget):
                                  )
 
 
+        #成功对话框
+        self.dialog = QDialog()
+        self.button = QPushButton('Shift跳过', self.dialog)
+        self.button.clicked.connect(self.dialog.close)
+        self.button.move(200, 50)
+        self.dialog.setWindowTitle('成功导入信息！！！')
+        #self.dialog.setWindowModality(Qt.ApplicationModal)
+
+
+
+
 
 
 
@@ -212,7 +223,12 @@ class Video(QWidget):
         file_name = self.model.fileName(index)
         file_path = self.model.filePath(index)
         global target_file_name
-        target_file_name=file_name
+
+        #target_file_name = file_name
+
+        #按文件路径命名
+        target_file_name=file_path
+
 
     def tree_cilcked(self, Qmodelidx):
 
@@ -514,6 +530,17 @@ class Video(QWidget):
 
 
 
+    # def Success_dialog(self):
+    #     dialog = QDialog()
+    #     button = QPushButton('确定', dialog)
+    #     button.clicked.connect(dialog.close)
+    #     button.move(50, 50)
+    #     dialog.setWindowTitle('对话框')
+    #     #dialog.setWindowModality(Qt.ApplicationModal)
+    #
+    #     dialog.exec()  # t弹出对话框
+
+
     def keyPressEvent(self, QKeyEvent):  # 键盘某个键被按下时调用
         # 参数1  控件
         if QKeyEvent.key() == Qt.Key_Left and self.lock and self.left<self.right and self.left_right_lock==False:
@@ -677,12 +704,18 @@ class Video(QWidget):
         global text_3
 
         global target_file_path
+        global target_file_name
 
+        global trigger
         if QKeyEvent.key() == Qt.Key_1 and self.lock:
             choice=1
+            trigger = False
 
         if QKeyEvent.key() == Qt.Key_2 and self.lock:
             choice=2
+            trigger=False
+
+
 
         if QKeyEvent.key() == Qt.Key_Space and self.lock:
 
@@ -715,18 +748,36 @@ class Video(QWidget):
 
 
         if QKeyEvent.key() == Qt.Key_M and self.lock:
+
+            target_file_name = target_file_name.replace('/', '$')
+            target_file_name = target_file_name.replace(':', '#')
             #先看文件是否已经存在了
             if os.path.exists(target_file_path+"/"+target_file_name[0:-4]+".txt"):
                 Warming = QMessageBox.warning(self, "Warming", "已经存在同名的文件了！！！",
                                                    QMessageBox.Yes)
             else:
-                f = open(target_file_path+"/"+target_file_name[0:-4]+".txt", 'w')
-                f.write(text_0+"\n"+text_1+"\n"+text_2+"\n"+text_3)
-                f.close()
+                #有的时候可能忘记标注某些帧，这个部分用于提醒，并且防止残缺的数据进入数据集
+                if (int(text_0.split(",")[0])==0 and int(text_0.split(",")[0])==0 ) or\
+                        (int(text_1.split(",")[0])==0 and int(text_1.split(",")[1])==0 and int(text_1.split(",")[2])==0 and int(text_1.split(",")[3])==0) or\
+                        (int(text_2.split(",")[0]) == 0 and int(text_2.split(",")[1]) == 0 and int(text_2.split(",")[2]) == 0 and int(text_2.split(",")[3]) == 0) or\
+                        (int(text_3.split(",")[0]) == 0 and int(text_3.split(",")[1]) == 0 and int(text_3.split(",")[2]) == 0 and int(text_3.split(",")[3]) == 0):
+                    Warming = QMessageBox.warning(self, "Warming", "您还有没有标注的帧！！！",
+                                                  QMessageBox.Yes)
+
+                else:
+                    f = open(target_file_path+"/"+target_file_name[0:-4]+".txt", 'w')
+                    f.write(text_0+"\n"+text_1+"\n"+text_2+"\n"+text_3)
+                    f.close()
 
             #这个函数就是为了：当我存储数据或者更换了新的视频源后刷新暂存的数据
             self.temp_clean()
 
+            #通过一个弹窗提示我已经成功输入数据，否则感觉没有反馈很不舒服
+            self.dialog.exec()  # t弹出对话框
+
+        #关闭成功对话框的按钮Enter
+        if QKeyEvent.key() == Qt.Key_Shift and self.lock:
+            self.dialog.close()
 
         #未来可能会用到的键组合
         if QKeyEvent.modifiers() == Qt.ControlModifier | Qt.ShiftModifier and QKeyEvent.key() == Qt.Key_A:  # 三键组合
@@ -750,13 +801,14 @@ pos_ball=[0,0,0,0]
 global clean_out
 clean_out=True
 
+global trigger
+trigger=True
+
 class myLabel(QLabel):
-    #这里的方框需要在我使用stop和load或者导入新的视频文件后消除，不能留下来
-    x0 = 0
-    y0 = 0
-    x1 = 0
-    y1 = 0
-    flag = False
+    # x0 = 0
+    # y0 = 0
+    # x1 = 0
+    # y1 = 0
     global choice
     choice=1
     human=QRect()
@@ -765,57 +817,240 @@ class myLabel(QLabel):
     global pos_human
     global pos_ball
 
+
+    flag = False  # 是否按下鼠标
+    ifmove = False  # 是否进行拖拽整体移动
+    biggerLeft = False  # 是否进行拖拽放大缩小，坐标点在标注框的左边
+    biggerRight = False  # 是否进行拖拽放大缩小，坐标点在标注框的右边
+    biggerButton = False  # 是否进行拖拽放大缩小，坐标点在标注框的下边
+    biggerTop = False  # 是否进行拖拽放大缩小，坐标点在标注框的上边
+    insurance = False# 上述图片中所涉及到的之外的位置
+    rect = []  # 标注框的起始点坐标和终点坐标，形如 [start_x, start_y, end_x, end_y]
+
+    # 鼠标双击事件
+    def mouseDoubleClickEvent(self, event):
+        self.rect = []
+        self.update()
+
+
+
+    # 鼠标单击事件
     def mousePressEvent(self, event):
         self.flag = True
-        self.x0 = event.x()
-        self.y0 = event.y()
+        # self.x0 = event.x()
+        # self.y0 = event.y()
 
+        global trigger
+        if self.rect and trigger:
+            # 此时已经存在标注框，再点击，则为拖拽整体移动或者放大缩小
+            # 整体移动：当点击坐标位于标注框内，则说明为整体移动
+            # 放大缩小：当点击坐标位于标注框的四条边上，则说明为放大缩小
+
+            # 框的左上角和右下角
+            topleft_x = min(self.rect[0], self.rect[2])
+            topleft_y = min(self.rect[1], self.rect[3])
+            buttonright_x = max(self.rect[0], self.rect[2])
+            buttonright_y = max(self.rect[1], self.rect[3])
+            # 整体移动的判定：点坐标在左上角和右下角之间
+            if (event.x() > topleft_x + 5) and (event.x() < buttonright_x - 5) and (event.y() > topleft_y + 5) and (
+                    event.y() < buttonright_y - 5):
+                print("判定为整体移动")
+                self.ifmove = True
+                self.preMousePosition = event.pos()
+            # 放大缩小的判定：点坐标在框的四条边上
+            # 上边，下边、左边、右边
+            # 当点坐标在标注框的左边时
+            if (event.x() > topleft_x - 5) and (event.x() < topleft_x + 5) and (event.y() > topleft_y) and (
+                    event.y() < buttonright_y):
+                print("判定为放大缩小，点坐标在标注框的左边上")
+                self.biggerLeft = True
+                self.preMousePosition = event.pos()
+            # 当点坐标在标注框的右边时
+            if (event.x() > buttonright_x - 5) and (event.x() < buttonright_x + 5) and (event.y() > topleft_y) and (
+                    event.y() < buttonright_y):
+                print("判定为放大缩小，点坐标在标注框的右边上")
+                self.biggerRight = True
+                self.preMousePosition = event.pos()
+            # 当点坐标在标注框的下边时
+            if (event.x() > topleft_x) and (event.x() < buttonright_x) and (event.y() > buttonright_y - 5) and (
+                    event.y() < buttonright_y + 5):
+                print("判定为放大缩小，点坐标在标注框的下边上")
+                self.biggerButton = True
+                self.preMousePosition = event.pos()
+            # 当点坐标在标注框的上边时
+            if (event.x() > topleft_x) and (event.x() < buttonright_x) and (event.y() > topleft_y - 5) and (
+                    event.y() < topleft_y + 5):
+                print("判定为放大缩小，点坐标在标注框的上边上")
+                self.biggerTop = True
+                self.preMousePosition = event.pos()
+            else:
+                #这个补充是为了防止我在拖动时离线太远导致上述所有的情况都没有发生
+                #然后就会执行花边框的操作，导致我只需要微调的框乱动
+                self.insurance = True
+
+        else:
+            trigger=True
+            # 如果当前没有标注框，则说明点击后开始拖拽生成标注框
+            # 此时，没有整体移动和放大缩小
+            self.ifmove = False
+            self.biggerLeft = False
+            self.biggerRight = False
+            self.biggerButton = False
+            self.biggerTop = False
+            self.insurance = False
+            self.rect = [event.x(), event.y(), event.x(), event.y()]
+
+    # 鼠标释放事件
     def mouseReleaseEvent(self, event):
         if choice==1:
-            self.human=QRect(self.x0, self.y0, self.x1 - self.x0, self.y1 - self.y0)
+            if self.rect:
+                self.human=QRect(self.rect[0],self.rect[1],self.rect[2]-self.rect[0],self.rect[3]-self.rect[1])
+            #self.human=QRect(self.x0, self.y0, self.x1 - self.x0, self.y1 - self.y0)
             pos_human[0]=self.human.x()
             pos_human[1]=self.human.y()
             pos_human[2] =self.human.x() + self.human.width()
             pos_human[3] =self.human.y() + self.human.height()
 
         if choice==2:
-            self.ball=QRect(self.x0, self.y0, self.x1 - self.x0, self.y1 - self.y0)
+            if self.rect:
+                self.ball = QRect(self.rect[0], self.rect[1], self.rect[2] - self.rect[0], self.rect[3] - self.rect[1])
+            #self.ball=QRect(self.x0, self.y0, self.x1 - self.x0, self.y1 - self.y0)
             pos_ball[0]=self.ball.x()
             pos_ball[1]=self.ball.y()
             pos_ball[2] =self.ball.x() + self.ball.width()
             pos_ball[3] =self.ball.y() + self.ball.height()
 
+
         self.flag = False
+        self.biggerLeft = False
+        self.biggerRight = False
+        self.biggerButton = False
+        self.biggerTop = False
+        self.insurance = False
+        self.ifmove = False
+
+        # 鼠标移动事件
 
     def mouseMoveEvent(self, event):
-        if self.flag:
-            self.x1 = event.x()
-            self.y1 = event.y()
+        if self.rect:
+            start_x, start_y = self.rect[0:2]
+
+        if self.flag:  # 按下鼠标，开始拖拽
+            # 拖拽可能有三种情况
+            # 情况1：在已有标注框的情况下，进行放大缩小
+            # 情况1.1：标注框的左边
+            if self.biggerLeft:
+                print("放大缩小：左边进行x轴方向的缩放")
+                differ = event.pos() - self.preMousePosition
+                # 更新标注框
+                start_x, start_y, end_x, end_y = self.rect
+                if start_x < end_x:
+                    start_x += differ.x()
+                else:
+                    end_x += differ.x()
+                self.rect = [start_x, start_y, end_x, end_y]
+                self.preMousePosition = event.pos()
+            # 情况1.2：标注框的右边
+            elif self.biggerRight:
+                print("放大缩小：右边进行x轴方向的缩放")
+                differ = event.pos() - self.preMousePosition
+                # 更新标注框
+                start_x, start_y, end_x, end_y = self.rect
+                if start_x > end_x:
+                    start_x += differ.x()
+                else:
+                    end_x += differ.x()
+                self.rect = [start_x, start_y, end_x, end_y]
+                self.preMousePosition = event.pos()
+            # 情况1.3：标注框的下边
+            elif self.biggerButton:
+                print("放大缩小：下边进行y轴方向的缩放")
+                differ = event.pos() - self.preMousePosition
+                # 更新标注框
+                start_x, start_y, end_x, end_y = self.rect
+                if start_y > end_y:
+                    start_y += differ.y()
+                else:
+                    end_y += differ.y()
+                self.rect = [start_x, start_y, end_x, end_y]
+                self.preMousePosition = event.pos()
+            # 情况1.4：标注框的上边
+            elif self.biggerTop:
+                print("放大缩小：上边进行y轴方向的缩放")
+                differ = event.pos() - self.preMousePosition
+                # 更新标注框
+                start_x, start_y, end_x, end_y = self.rect
+                if start_y < end_y:
+                    start_y += differ.y()
+                else:
+                    end_y += differ.y()
+                self.rect = [start_x, start_y, end_x, end_y]
+                self.preMousePosition = event.pos()
+
+            # 情况2：在已有标注框的情况下，进行整体移动
+            elif self.ifmove:
+                # differ = event.pos() - self.preMousePosition
+                # # 更新标注框
+                # start_x, start_y, end_x, end_y = self.rect
+                # start_x, start_y = start_x + differ.x(), start_y + differ.y()
+                # end_x, end_y = end_x + differ.x(), end_y + differ.y()
+                # self.rect = [start_x, start_y, end_x, end_y]
+                # self.preMousePosition = event.pos()
+                print("整体移动")
+            elif self.insurance:
+                print("nothing will happen")
+            # 情况3：做标注框
+            else:
+                # self.x1 = event.x()
+                # self.y1 = event.y()
+                self.rect = [start_x, start_y, event.x(), event.y()]
+
             self.update()
 
+    # 绘制事件
     def paintEvent(self, event):
-        super().paintEvent(event)
-        #这个rect是为了实现画rect的动画，让我画框的动作连贯
-        rect = QRect(self.x0, self.y0, self.x1 - self.x0, self.y1 - self.y0)
-        painter = QPainter(self)
-
         #导入新视频或者stop或者load后，清空方框
         global clean_out
         if clean_out==False:
+            #print("cheee")
+            self.rect=[]
+            self.update()
+
             self.human=QRect()
             self.ball=QRect()
-            rect=QRect()
+            real_rect=QRect()
             #painter.drawRect(rect)
-
             #每次更换后，choice自动变为1，成为human
             global choice
             choice=1
             clean_out=True
-
             global pos_human
             global pos_ball
             pos_human=[0,0,0,0]
             pos_ball=[0,0,0,0]
+
+
+        super().paintEvent(event)
+        if len(self.rect) == 0: return
+        # rect =QRect(self.x0, self.y0, abs(self.x1-self.x0), abs(self.y1-self.y0))
+        # 这里换成任意方向的矩形框
+        # 标注框的左上角和右下角坐标
+        x0=self.rect[0]
+        y0 = self.rect[1]
+        x1 = self.rect[2]
+        y1 = self.rect[3]
+        # x0 = min(self.rect[0], self.rect[2])
+        # y0 = min(self.rect[1], self.rect[3])
+        # x1 = max(self.rect[0], self.rect[2])
+        # y1 = max(self.rect[1], self.rect[3])
+        width = x1 - x0
+        height = y1 - y0
+
+        # 矩形
+        real_rect = QRect(x0, y0, width, height)
+        painter = QPainter(self)
+        #painter.setPen(QPen(Qt.red, 2, Qt.SolidLine))
+
 
 
         if choice==1:
@@ -830,11 +1065,21 @@ class myLabel(QLabel):
 
             painter.setPen(QPen(Qt.red, self.line_width, Qt.SolidLine))
         #让我画框的动作连贯
-        painter.drawRect(rect)
+
+        painter.drawRect(real_rect)
+
+
 
 
 if __name__ == "__main__":
     app = QtWidgets.QApplication(sys.argv)
     my = Video()
+
+
+    #修改界面的背景颜色
+    my.setStyleSheet('''QWidget{background-color:silver;}''')
+    #可以选定自己喜欢的背景图片
+    #my.setStyleSheet("#x{border-image:url(x.png)}")
+
     my.show()
     sys.exit(app.exec_())
